+++ packages/browser/index.js
@@ -1,2 +1,2 @@
 export { mountBrowser } from './src/browser.js';
-export const BROWSER_VERSION = '3.2.3';
+export const BROWSER_VERSION = '3.3.0';